+++ CHANGES.rst
@@ -4,7 +4,8 @@ History
 0.2.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add versions command
+  [kakshay21]
 
 
 0.2.0 (2018-04-03)